Move PrivateRoute to containers
