+++ src/containers/Routes.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route } from 'react-router-dom';
 
-import PrivateRoute from '../components/PrivateRoute';
+import PrivateRoute from './PrivateRoute';
 import Booking from '../components/Booking';
 import Dashboard from '../components/Dashboard';
 import MasterForm from '../components/registration/MasterForm';
